chore(fake-signalr): configure to use @fakehost/fake-rest
