+++ package.json
@@ -22,5 +22,8 @@
     "build": "lerna run --stream build",
     "lint": "lerna run --stream lint --",
     "test": "lerna run --stream test --"
+  },
+  "resolutions": {
+    "whatwg-url": "12.0.1"
   }
 }
