+++ packages/signalr/fake-signalr/package.json
@@ -19,8 +19,10 @@
     "license": "MIT",
     "dependencies": {
         "@fakehost/exchange": "^0.2.0",
+        "@fakehost/fake-rest": "^1.0.0",
         "@fakehost/host": "^1.0.0",
         "uuid": "^9.0.0",
+        "url": "^0.11.0",
         "ws": "^8.13.0"
     },
     "devDependencies": {
